Another attempt at making MIPS happy
(I really don't understand how arithmetic makes it segfault...)
